refactor: overhaul services in shared
