+++ packages/shared/src/services/NotificationService.ts
@@ -1,11 +1,17 @@
 import { Schema as S } from '@effect/schema';
 import type { Effect } from 'effect';
 import { Context } from 'effect';
-import { BaseError } from '../index.js';
 
 export const notificationOptionsSchema = S.Struct({
 	id: S.optional(S.String),
-	...BaseError.fields,
+	title: S.String,
+	description: S.String,
+	action: S.optional(
+		S.Struct({
+			label: S.String,
+			goto: S.String,
+		}),
+	),
 });
 
 type NotificationOptions = S.Schema.Type<typeof notificationOptionsSchema>;
