+++ packages/shared/src/services/index.ts
@@ -1,3 +1,4 @@
 export * from './NotificationService.js';
+export * from './ToastService.js';
 export * from './TranscriptionService.js';
 export * from './TranscriptionServiceWhisperingLive.js';
